refactor(edges): extract pathing into hook
